requirements.txt
streamlit
pyautogui
openai
requests
logging
python-dotenv
pytest